requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pandas>=1.5.0
numpy>=1.21.0
opencv-python>=4.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
gradio>=3.50.0
reportlab>=4.0.0
tqdm>=4.65.0
pillow>=9.0.0
albumentations>=1.3.0
efficientnet-pytorch>=0.7.1